+++ syft/core/frameworks/torch/__init__.py
@@ -1,10 +1,11 @@
 from .hook import TorchHook
 from .tensor import _SyftTensor, _LocalTensor, _PointerTensor
-from .tensor import _FixedPrecisionTensor, _TorchTensor, _PlusIsMinusTensor, _MPCTensor
+from .tensor import _FixedPrecisionTensor, _TorchTensor, _PlusIsMinusTensor,_GeneralizedPointerTensor
+from .mpc_tensor import _MPCTensor
 
 __all__ = ['TorchHook', '_SyftTensor', '_LocalTensor',
            '_PointerTensor', '_FixedPrecisionTensor', '_TorchTensor',
-           '_PlusIsMinusTensor', '_MPCTensor', 'encode']
+           '_PlusIsMinusTensor','_GeneralizedPointerTensor', '_MPCTensor', 'encode']
 
 import torch
