+++ syft/core/frameworks/torch/hook.py
@@ -1,4 +1,4 @@
-dmport torch
+import torch
 import random
 import inspect
 import re
